Day 13 part 2
Unsure I would have solved this without the tip that the system of
congruences modulo the bus numbers is a description of the chinese
remainder theorem.

The other hard part of this, that I struggled with for at least another
hour was to realize that the remainders are not the offset from the bus
number to the first timestamp, but rather that offset minus the bus
number.