Day 19 part 2
Ouch that was difficult. Figured out fairly quickly that I should try to
bulid a regular expressiion to match these, but acciidentally shot
myself in the foot by putting in the dupliciate rules from the
instructions far away from the code I was trying to write. Lost about an
hour or more debugging that joy.

And then I didn't realize that the rule 11 needs to be balanced. I tried
to use a+b+ which doesn't work since they obviously need to be the same
length. The regex I created in that case is a huge hack, and would fail
for some known inputs.