Day 20 part 2 WIP
Have so far placed two tiles successfully.

Initially my strategy was to try to build up the entire puzzle at once
but that did not seem tenable. Instead, after much deliberation, I
decided that building up the border which has the constraint of forcing
some of the edges to only appear once within the tiles will make further
solving the puzzle easier.

Once we have the border, we can consider the interior, which has the
contraint of two edges needing to be matching, which should force the
one correct orientation for the remaining puzzle pieces.